patch-build-shim
I added `return nil` after `prepareShim` as I was working with Azure Trigger.I was just concerned with building the Docker image and not with the binary itself. Since I was not working with binary, this error went unnoticed. This only came into light when I was messing with cli trigger. We might need to discuss whether we want to build the binaries of the app, which only needs other artifacts.